requirements.txt

# snailjob
snail-job-python==0.0.2
